+++ reikna/cluda/cuda.py
@@ -106,7 +106,10 @@ class Array(gpuarray.GPUArray):
         so we're overriding it.
         """
         new_arr = self._new_like_me()
-        gpuarray._memcpy_discontig(new_arr, self, async_=True, stream=self.thread._queue)
+        # FIXME: a temporary workaround for PyCUDA not being compatible with Py3.7
+        # where `async` is a keyword.
+        kwds = {'async': True, 'stream': self.thread._queue}
+        gpuarray._memcpy_discontig(new_arr, self, **kwds)
         return new_arr
 
     def _new_like_me(self, dtype=None):